+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [3.67.28]
+
+## Fixed
+
+- chatbot apis now supports async botFactory
+
+# [3.67.27]
+
+## Fixed
+
+- reporting now uses an updated state instead of the initial
+
 # [3.67.20-6]
 
 ## Fixed